Create ServiceSniffer.py
ServiceSniffer is a python script used to sniff local services running on your server
It is doing this by monitoring server TCP sessions and count the number of packets. 
A timestamp, TCP port and number of packets are stored in an XML file every 59s by default.
You can utilize these data to pull it and store it in a DB to be displayed in a graph and monitored by NOC team.